Fix erroneous constant loading of additional rows
The issue I needed to resolve was that at some point null cursor was returned and then the table was reloading rows from the start all over again and again (when the last row was visible)

More correct implementation would have been to constantly return the "next" record. But in this case we didn't have any data for it because "next" pointed to a record that isn't there.

anyway, right now we have lost the option of loading log records that are yet to be added to the table. But this at least fixes infinite reload of the entire table.